requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.11.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
exceptiongroup==1.1.1
flake8==6.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
PySocks==1.7.1
pytest==7.3.1
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0
python-dotenv==1.0.0
